feat(RichTextEditor): lazy load quill
