+++ .changeset/breezy-cheetahs-guess.md
@@ -1,0 +1,10 @@
+---
+'@toptal/picasso': minor
+---
+
+---
+
+### RichTextEditor
+
+- quill editor is lazy-loaded using React.lazy and React.Suspense
+- [quill](https://www.npmjs.com/package/quill) is no longer part of main bundle
