+++ packages/picasso/src/QuillEditor/index.ts
@@ -1,4 +1,4 @@
-export { default } from './QuillEditor'
+export { default } from './LazyQuillEditor'
 export type { Props as QuillEditorProps } from './QuillEditor'
 export * from './types'
 export { CUSTOM_QUILL_EDITOR_FORMAT_EVENT } from './constants'
