+++ packages/picasso/src/QuillEditorOutput/index.ts
@@ -1,0 +1,1 @@
+export { default } from './QuillEditorOutput'
